Move the `StateStructure.pull` test to `StateStructureDenotationalUnitTests`
