+++ denotational/core/src/test/kotlin/io/github/oxidefrp/core/StateStructureSharedUnitTests.kt
@@ -56,75 +56,6 @@ object StateStructureSharedUnitTests {
         }
     }
 
-    object Pull {
-        @Test
-        fun test() = testSystem {
-            val stateSignal = buildInputSignal {
-                when (it) {
-                    Tick(t = 30) -> S(sum = 1)
-                    else -> throw UnsupportedOperationException("Unexpected tick: $it")
-                }
-            }
-
-            val inputStateStream = buildInputStream(
-                EventOccurrenceDesc(tick = Tick(t = 10), event = S(sum = 10)),
-                EventOccurrenceDesc(tick = Tick(t = 25), event = S(sum = 25)),
-                EventOccurrenceDesc(tick = Tick(t = 40), event = S(sum = 40)),
-                EventOccurrenceDesc(tick = Tick(t = 50), event = S(sum = 50)),
-            )
-
-            val inputLayer = StateSchedulerLayer(stateStream = inputStateStream)
-
-            val extraStream = buildInputStream(
-                EventOccurrenceDesc(tick = Tick(t = 20), event = S(sum = 20)),
-                EventOccurrenceDesc(tick = Tick(t = 45), event = S(sum = 45)),
-            )
-
-            val momentStateStructure = stateStructure(
-                n = 2,
-                extra = extraStream,
-                result = {
-                    object : Moment<String>() {
-                        override fun pullDirectly(t: Time): String = "${it.sum}@${t.t}"
-                    }
-                },
-            )
-
-            val (outputLayer, resultValue) = StateStructure.pull(momentStateStructure).constructDirectly(
-                stateSignal = stateSignal,
-            ).pullEnterDirectly(
-                t = Time(t = 30.0),
-                oldState = inputLayer,
-            )
-
-            TestSpec(
-                checks = listOf(
-                    TestCheck(
-                        subject = outputLayer.stateStream,
-                        name = "Output state stream",
-                        spec = EventStreamSpec(
-                            expectedEvents = listOf(
-                                EventOccurrenceDesc(tick = Tick(10), event = S(sum = 12)),
-                                EventOccurrenceDesc(tick = Tick(20), event = S(sum = 20)),
-                                EventOccurrenceDesc(tick = Tick(25), event = S(sum = 27)),
-                                EventOccurrenceDesc(tick = Tick(40), event = S(sum = 42)),
-                                EventOccurrenceDesc(tick = Tick(45), event = S(sum = 45)),
-                                EventOccurrenceDesc(tick = Tick(50), event = S(sum = 52)),
-                            )
-                        ),
-                    ),
-                    TestCheck(
-                        subject = resultValue,
-                        name = "Result value",
-                        spec = ValueSpec(
-                            expected = "1@30.0",
-                        ),
-                    ),
-                ),
-            )
-        }
-    }
-
     object CellConstruct {
         @Test
         fun test() = testSystem {
